fix: menuとtask追加ボタンの表示非表示ロジックが似ていたので共通化
